requirements.txt
asgiref==3.5.0
branca==0.4.2
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
folium==0.12.1.post1
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.1.1
numpy==1.22.3
Pillow==9.0.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
